Create simpleContract_PseudoCode.js
This simple smart contract defines a `SimpleContract` with a single variable to store a message.

 The contract also defines an event to indicate when the message has been changed, as well as functions to set and get the message.

The `setMessage` function allows anyone to set the message to a new value by calling the function with a string parameter.
 The function sets the message to the new value and emits the MessageChanged event to indicate that the message has been changed.

The `getMessage` function allows anyone to retrieve the current value of the message by calling the function with no parameters. 
The function simply returns the value of the message variable.